backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
reportlab==4.0.0
psycopg2-binary==2.9.10
alembic==1.12.0